Simplify: extract shared icon-set utilities, cache getSets()
- Add src/lib/icon-sets/utils.ts with toPascalCase, createImgComponent,
  encodeSvgToDataUrl, and isForwardRefComponent — eliminates duplication
  across 4-5 icon-set files
- toPascalCase: was copied in lucide, simple-icons, phosphor, tabler
- createImgComponent: was copied in aws, devicon, icons.ts
- encodeSvgToDataUrl: btoa/base64 pattern was copied in lucide, simple-icons,
  phosphor, tabler
- isForwardRefComponent + REACT_FORWARD_REF: was duplicated in phosphor, tabler
- Cache getSets() result in IconRegistry; invalidate on register() so all
  future calls return the same array reference without re-sorting
