+++ src/lib/icon-sets/aws.ts
@@ -1,6 +1,6 @@
-import { createElement } from 'react'
 import { iconRegistry } from '@/lib/icon-registry'
 import type { IconComponent, IconEntry } from '@/lib/icon-registry'
+import { createImgComponent } from '@/lib/icon-sets/utils'
 
 const awsArchGlob = import.meta.glob<string>(
   '/node_modules/aws-svg-icons/lib/Architecture-Service-Icons_07302021/*/64/*.svg',
@@ -42,24 +42,6 @@ const awsIconNames = new Set(Object.keys(awsIndex))
 export const AWS_ICONS: IconEntry[] = Object.entries(awsIndex)
   .map(([key, entry]) => ({ label: entry.label, value: key }))
   .sort((a, b) => a.label.localeCompare(b.label))
-
-function createImgComponent(url: string): IconComponent {
-  const ImgIcon: IconComponent = (props) => {
-    const { className, width, height, ...rest } = props as Record<
-      string,
-      unknown
-    >
-    return createElement('img', {
-      src: url,
-      alt: '',
-      className,
-      width: width ?? 16,
-      height: height ?? 16,
-      ...rest,
-    })
-  }
-  return ImgIcon
-}
 
 function resolveAwsUrl(iconName: string): string | null {
   const key = iconName.toLowerCase()
