+++ src/lib/icon-sets/devicon.ts
@@ -1,6 +1,6 @@
-import { createElement } from 'react'
 import { iconRegistry } from '@/lib/icon-registry'
 import type { IconComponent, IconEntry } from '@/lib/icon-registry'
+import { createImgComponent } from '@/lib/icon-sets/utils'
 
 const deviconGlob = import.meta.glob<string>(
   '/node_modules/devicon/icons/**/*.svg',
@@ -21,24 +21,6 @@ export const DEVICON_ICONS: IconEntry[] = Object.entries(deviconIndex)
     value,
   }))
   .sort((a, b) => a.label.localeCompare(b.label))
-
-function createImgComponent(url: string): IconComponent {
-  const ImgIcon: IconComponent = (props) => {
-    const { className, width, height, ...rest } = props as Record<
-      string,
-      unknown
-    >
-    return createElement('img', {
-      src: url,
-      alt: '',
-      className,
-      width: width ?? 16,
-      height: height ?? 16,
-      ...rest,
-    })
-  }
-  return ImgIcon
-}
 
 function resolve(value: string): IconComponent | null {
   if (!value.startsWith('devicon-')) return null
